# docs/swagger.yaml
basePath: /api
definitions:
  budget.CreateBudgetRequest:
    properties:
      budget:
        example: "5000000.00"
        type: string
      date:
        example: "2025-01-01"
        type: string
    required:
    - budget
    - date
    type: object
  budget.MonthlyBudget:
    properties:
      budget:
        example: 5000000
        type: number
      date:
        example: "2025-01-01T00:00:00Z"
        type: string
      id:
        example: bf8a39e8-4226-4d04-a035-6453181878d6
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  budget.UpdateBudgetRequest:
    properties:
      budget:
        example: "5000000.00"
        type: string
      date:
        example: "2025-01-01"
        type: string
    required:
    - budget
    - date
    type: object
  transaction.CreateTransactionRequest:
    properties:
      amount:
        example: "50000"
        type: string
      budget_id:
        example: bf8a39e8-4226-4d04-a035-6453181878d6
        type: string
      date:
        example: "2025-01-02"
        type: string
    required:
    - amount
    - budget_id
    - date
    type: object
  transaction.Transaction:
    properties:
      amount:
        example: 300000
        type: number
      budget_id:
        example: bf8a39e8-4226-4d04-a035-6453181878d6
        type: string
      date:
        example: "2025-01-01T00:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  transaction.UpdateTransactionRequest:
    properties:
      amount:
        example: "75000"
        type: string
      date:
        example: "2025-01-03"
        type: string
    required:
    - amount
    - date
    type: object
  user.LoginRequest:
    properties:
      email:
        example: tubagus@example.com
        type: string
      password:
        example: RahasiaNegara123!
        type: string
    required:
    - email
    - password
    type: object
  user.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      token_type:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/user.UserResponse'
    type: object
  user.RegisterRequest:
    properties:
      email:
        example: tubagus@example.com
        type: string
      password:
        example: RahasiaNegara123!
        minLength: 6
        type: string
      username:
        example: tubagus_aldi
        maxLength: 30
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  user.RegisterResponse:
    properties:
      created_at:
        example: "2025-01-01T15:00:00Z"
        type: string
      email:
        example: tubagus@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      username:
        example: tubagus_aldi
        type: string
    type: object
  user.UserResponse:
    properties:
      created_at:
        example: "2025-01-01T15:00:00Z"
        type: string
      email:
        example: tubagus@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      username:
        example: tubagus_aldi
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: admin@tubsamy.tech
    name: Tubagus Aldi
  description: API Documentation for Finance Tracker App.
  termsOfService: http://swagger.io/terms/
  title: Finance Tracker API
  version: "1.0"
paths:
  /budgets:
    get:
      consumes:
      - application/json
      description: Get list of monthly budgets with date filter
      parameters:
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/budget.MonthlyBudget'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Budgets
      tags:
      - Budget
    post:
      consumes:
      - application/json
      description: Create a new global budget for a specific month
      parameters:
      - description: Budget Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/budget.CreateBudgetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/budget.MonthlyBudget'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create Monthly Budget
      tags:
      - Budget
  /budgets/{id}:
    put:
      consumes:
      - application/json
      description: Update existing budget amount or date
      parameters:
      - description: Budget ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/budget.UpdateBudgetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/budget.MonthlyBudget'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Monthly Budget
      tags:
      - Budget
  /transactions:
    get:
      consumes:
      - application/json
      description: Get list of expenses for a specific account
      parameters:
      - description: Account ID (UUID)
        in: query
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/transaction.Transaction'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Transactions
      tags:
      - Transaction
    post:
      consumes:
      - application/json
      description: Record a new expense for a specific account
      parameters:
      - description: Transaction Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/transaction.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/transaction.Transaction'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create Transaction
      tags:
      - Transaction
  /transactions/{id}:
    put:
      consumes:
      - application/json
      description: Update existing expense record
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/transaction.UpdateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transaction.Transaction'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Transaction
      tags:
      - Transaction
  /users/current:
    get:
      consumes:
      - application/json
      description: Get currently logged in user profile based on JWT Token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Current User Profile
      tags:
      - User
  /users/login:
    post:
      consumes:
      - application/json
      description: Login and get JWT Token
      parameters:
      - description: Login Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login User
      tags:
      - User
  /users/logout:
    post:
      consumes:
      - application/json
      description: Invalidate current session (client-side token deletion recommended)
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Logout successful'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout User
      tags:
      - User
  /users/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Register Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register new user
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"